[Fix] 친구 삭제 - JSON 예외처리할 때 생기는 에러

+++ src/main/java/com/uplus/bugzerobackend/service/FriendService.java
@@ -8,8 +8,9 @@ import com.uplus.bugzerobackend.dto.FriendListDto;
 import com.uplus.bugzerobackend.mapper.FriendMapper;
 import com.uplus.bugzerobackend.mapper.UserMapper;
 
-import org.springframework.stereotype.Service;
+import jakarta.persistence.EntityNotFoundException;
 
+import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -20,13 +21,12 @@ public class FriendService {
     private final FriendMapper friendMapper;
     private final UserMapper userMapper;
     private final ObjectMapper objectMapper = new ObjectMapper();
-
-
+    
     public FriendService(FriendMapper friendMapper, UserMapper userMapper) {
         this.friendMapper = friendMapper;
         this.userMapper = userMapper;
     }
-
+    
     public List<FriendListDto> getFriend(Integer userId) {
         Friend friend = friendMapper.getFriendById(userId);
         if (friend == null || friend.getFriendList() == null) {
@@ -34,7 +34,6 @@ public class FriendService {
         }
         List<Integer> friendIds = friend.getFriendList();
         List<FriendListDto> response = new ArrayList<>();
-
         for(Integer friendId : friendIds) {
             User user = userMapper.getUserById(friendId);
             if(user!=null) {
@@ -43,7 +42,6 @@ public class FriendService {
         }
         return response;
     }
-    
 	public void deleteFriend(int userId, int friendId) {
 		Map<String, Object> rParams = new HashMap<>();
 		rParams.put("userId", userId);
@@ -52,7 +50,11 @@ public class FriendService {
 		Map<String, Object> sParams = new HashMap<>();
 		sParams.put("userId", friendId);
 		sParams.put("friendId", userId);
-	    
+		
+		if(friendMapper.checkFriendExistence(rParams) == 0) {
+			throw new EntityNotFoundException("존재하지 않는 친구입니다.");
+		}
+	   
 	    friendMapper.deleteFriend(rParams);
 	    friendMapper.deleteFriend(sParams);
 	}
